Fix the capitalization problem of Describe Site Monitor Data.

+++ services/cms/describe_site_monitor_statistics.go
@@ -89,7 +89,7 @@ type DescribeSiteMonitorStatisticsResponse struct {
 	Message   string `json:"Message" xml:"Message"`
 	Success   string `json:"Success" xml:"Success"`
 	RequestId string `json:"RequestId" xml:"RequestId"`
-	Data      int    `json:"Data" xml:"Data"`
+	Data      int64  `json:"Data" xml:"Data"`
 }
 
 // CreateDescribeSiteMonitorStatisticsRequest creates a request to invoke DescribeSiteMonitorStatistics API
