+++ services/cms/struct_node_task_config.go
@@ -17,12 +17,12 @@ package cms
 
 // NodeTaskConfig is a nested struct in cms response
 type NodeTaskConfig struct {
-	Id          int         `json:"Id" xml:"Id"`
+	Id          int64       `json:"Id" xml:"Id"`
 	TaskName    string      `json:"TaskName" xml:"TaskName"`
 	TaskType    string      `json:"TaskType" xml:"TaskType"`
 	TaskScope   string      `json:"TaskScope" xml:"TaskScope"`
 	Disabled    bool        `json:"Disabled" xml:"Disabled"`
-	GroupId     int         `json:"GroupId" xml:"GroupId"`
+	GroupId     int64       `json:"GroupId" xml:"GroupId"`
 	GroupName   string      `json:"GroupName" xml:"GroupName"`
 	Instances   Instances   `json:"Instances" xml:"Instances"`
 	TaskOption  TaskOption  `json:"TaskOption" xml:"TaskOption"`
