+++ services/cms/struct_notify_policy.go
@@ -19,7 +19,7 @@ package cms
 type NotifyPolicy struct {
 	Type      string `json:"Type" xml:"Type"`
 	Id        string `json:"Id" xml:"Id"`
-	StartTime int    `json:"StartTime" xml:"StartTime"`
-	EndTime   int    `json:"EndTime" xml:"EndTime"`
+	StartTime int64  `json:"StartTime" xml:"StartTime"`
+	EndTime   int64  `json:"EndTime" xml:"EndTime"`
 	GroupId   string `json:"GroupId" xml:"GroupId"`
 }
