+++ services/cms/struct_resource_item.go
@@ -17,7 +17,7 @@ package cms
 
 // ResourceItem is a nested struct in cms response
 type ResourceItem struct {
-	GroupId        int     `json:"GroupId" xml:"GroupId"`
+	GroupId        int64   `json:"GroupId" xml:"GroupId"`
 	Category       string  `json:"Category" xml:"Category"`
 	FilterRelation string  `json:"FilterRelation" xml:"FilterRelation"`
 	Filters        Filters `json:"Filters" xml:"Filters"`
